Style - Empty Methods
Methods with an empty implementation add to the project's overhead because the assumption quickly becomes that those methods are implemented (as a noop) for a reason.

Apple's template code is notorious for for leaving these turds all over the place! Empty methods or methods that only call through to super shouldn't make it past code review.